Split <Root /> into subcomponents
