+++ frontend/src/pages/Root/_drawer.scss
@@ -1,0 +1,2 @@
+$drawer-width: 240px;
+$drawer-height: 64px;
