+++ frontend/src/pages/Root/shared.module.scss
@@ -1,0 +1,9 @@
+@use "drawer";
+
+.drawerHeader {
+  display: flex;
+  align-items: center;
+  padding: 0 8px;
+  justify-content: flex-end;
+  min-height: drawer.$drawer-height;
+}
